Move NegotiateToResponse under ToResponse.Negotiable
