+++ core/src/main/scala/io/finch/ToResponse.scala
@@ -89,6 +89,41 @@ trait HighPriorityToResponseInstances extends LowPriorityToResponseInstances {
 
 object ToResponse extends HighPriorityToResponseInstances {
 
+  /**
+   * Enables server-driven content negotiation with client.
+   *
+   * Picks corresponding instance of `ToResponse` according to `Accept` header of a request
+   */
+  trait Negotiable[A, CT] {
+    def apply(accept: Seq[Accept]): ToResponse.Aux[A, CT]
+  }
+
+  object Negotiable {
+
+    implicit def coproductToNegotiable[A, CTH <: String, CTT <: Coproduct](implicit
+      h: ToResponse.Aux[A, CTH],
+      t: Negotiable[A, CTT],
+      a: Accept.Matcher[CTH]
+    ): Negotiable[A, CTH :+: CTT] = new Negotiable[A, CTH :+: CTT] {
+      def apply(accept: Seq[Accept]): ToResponse.Aux[A, CTH :+: CTT] =
+        if (accept.exists(_.matches[CTH])) h.asInstanceOf[ToResponse.Aux[A, CTH :+: CTT]]
+        else t(accept).asInstanceOf[ToResponse.Aux[A, CTH :+: CTT]]
+    }
+
+    implicit def cnilToNegotiable[A, CTH <: String](implicit
+      tr: ToResponse.Aux[A, CTH]
+    ): Negotiable[A, CTH :+: CNil] = new Negotiable[A, CTH :+: CNil] {
+      def apply(accept: Seq[Accept]): ToResponse.Aux[A, CTH :+: CNil] =
+        tr.asInstanceOf[ToResponse.Aux[A, CTH :+: CNil]]
+    }
+
+    implicit def singleToNegotiable[A, CT <: String](implicit
+      tr: ToResponse.Aux[A, CT]
+    ): Negotiable[A, CT] = new Negotiable[A, CT] {
+      def apply(accept: Seq[Accept]): ToResponse.Aux[A, CT] = tr
+    }
+  }
+
   implicit def cnilToResponse[CT <: String]: Aux[CNil, CT] =
     instance((_, _) => Response(Version.Http10, Status.NotFound))
